User creation on production vagrant (fixes #3869) (#3874)

+++ docker/planet/nginx/credentials.sh
@@ -3,13 +3,19 @@ echo "HTTP/1.0 200 OK"
 echo "Content-type: text/plain"
 echo ""
 
+YML_PATH=credentials/credentials.yml
 PLANET_USER=${PLANET_CREDENTIALS%%,*}
 PLANET_PASS=${PLANET_CREDENTIALS##*,}
+if [ -f "$YML_PATH" ]; then
+  OLD_USER=$(grep COUCHDB_USER $YML_PATH | sed -e 's/.*=//')
+fi
 
-rm credentials/credentials.yml
-echo "services:
-  db-init:
-    environment:
-      - COUCHDB_USER=$PLANET_USER
-      - COUCHDB_PASS=$PLANET_PASS
-version: \"2\"" >> credentials/credentials.yml
+if [ -z "$OLD_USER" ] || [ "$PLANET_USER" == "$OLD_USER" ]; then
+  rm credentials/credentials.yml
+  echo "services:
+    db-init:
+      environment:
+        - COUCHDB_USER=$PLANET_USER
+        - COUCHDB_PASS=$PLANET_PASS
+  version: \"2\"" >> $YML_PATH
+fi
